Finished disassembly of server.js into user.js and conversation.js.
