+++ server/src/server.js
@@ -61,19 +61,14 @@ const pgQueryP_readOnly = pg.pgQueryP_readOnly;
 const pgQueryP_readOnly_wRetryIfEmpty = pg.pgQueryP_readOnly_wRetryIfEmpty;
 
 const User = require('./user');
-const pidCache = User.pidCache;
-const getPidPromise = User.getPidPromise;
 
 const Conversation = require('./conversation');
-const getZidFromConversationId = Conversation.getZidFromConversationId;
 
 // # Slack setup
 
 var WebClient = require('@slack/client').WebClient;
 var web = new WebClient(process.env.SLACK_API_TOKEN);
 // const winston = require("winston");
-
-
 // # notifications
 const winston = console;
 const emailSenders = require('./email/sendEmailSesMailgun').EmailSenders(AWS);
@@ -99,8 +94,6 @@ if (useTranslateApi) {
     projectId: JSON.parse(fs.readFileSync(GOOGLE_CREDS_TEMP_FILENAME)).project_id,
   });
 }
-
-
 //var SegfaultHandler = require('segfault-handler');
  
 //SegfaultHandler.registerHandler("segfault.log");
@@ -116,9 +109,6 @@ if (useTranslateApi) {
 //   console.error(5);
 //   console.error(err);
 // });
-
-
-
 
 if (devMode) {
   Promise.longStackTraces();
@@ -150,29 +140,16 @@ Promise.onPossiblyUnhandledRejection(function(err) {
   // throw err; // not throwing since we're printing stack traces anyway
 });
 
-
-
 const adminEmailDataExport = process.env.ADMIN_EMAIL_DATA_EXPORT || ""
 const adminEmailDataExportTest = process.env.ADMIN_EMAIL_DATA_EXPORT_TEST || ""
 const adminEmailEmailTest = process.env.ADMIN_EMAIL_EMAIL_TEST || ""
 
 const admin_emails = process.env.ADMIN_EMAILS ? JSON.parse(process.env.ADMIN_EMAILS) : [];
 const polisDevs = process.env.ADMIN_UIDS ? JSON.parse(process.env.ADMIN_UIDS) : [];
-
-
 function isPolisDev(uid) {
   console.log("polisDevs", polisDevs)
   return polisDevs.indexOf(uid) >= 0;
 }
-
-// so we can grant extra days to users
-// eventually we should probably move this to db.
-// for now, use git blame to see when these were added
-const  usersToAdditionalTrialDays = {
-  50756: 14, // julien
-  85423: 100, // mike test
-};
-
 // log heap stats
 setInterval(function() {
   let mem = process.memoryUsage();
@@ -186,8 +163,6 @@ setInterval(function() {
   //metric("api.process.mem.rss", rss, start);
   //metric("api.process.mem.heapTotal", heapTotal, start);
 }, 10 * 1000);
-
-
 // // BEGIN GITHUB OAUTH2
 // let CLIENT_SECRET = "0b178e412a10fa023a0153bf7cefaf6dae0f74b9";
 // let CLIENT_ID = "109a1eb4732b3ec1075b";
@@ -207,8 +182,6 @@ setInterval(function() {
 //   state: '3(#0/!~'
 // });
 // // END GITHUB OAUTH2
-
-
 const POLIS_FROM_ADDRESS = process.env.POLIS_FROM_ADDRESS;
 
 const akismet = akismetLib.client({
@@ -223,18 +196,11 @@ akismet.verifyKey(function(err, verified) {
     winston.log("info", 'Akismet: Unable to verify API key.');
   }
 });
-
-
 // let SELF_HOSTNAME = "localhost:" + process.env.PORT;
 // if (!devMode) {
 // ^^^ possible to use localhost on Heroku?
 //  SELF_HOSTNAME = process.env.SERVICE_HOSTNAME
 //}
-
-
-
-
-
 
 function isSpam(o) {
   return new MPromise("isSpam", function(resolve, reject) {
@@ -258,8 +224,6 @@ if (devMode) {
 } else {
   INFO = function() {};
 }
-
-
 
 // basic defaultdict implementation
 function DD(f) {
@@ -286,8 +250,6 @@ DD.prototype.s = DA.prototype.s = function(k, v) {
 //   return [];
 // }
 
-
-
 const domainOverride = process.env.DOMAIN_OVERRIDE || null;
 
 function haltOnTimeout(req, res, next) {
@@ -305,8 +267,6 @@ function ifDefinedSet(name, source, dest) {
 }
 
 //metric("api.process.launch", 1);
-
-
 // TODO clean this up
 // const intercom = new Intercom(process.env.INTERCOM_ACCESS_TOKEN);
 const intercom = intercomClient;
@@ -461,8 +421,6 @@ const sql_reports = sql.define({
   ],
 });
 
-
-
 // // Eventually, the plan is to support a larger number-space by using some lowercase letters.
 // // Waiting to implement that since there's cognitive overhead with mapping the IDs to/from
 // // letters/numbers.
@@ -495,11 +453,6 @@ const sql_reports = sql.define({
 //     };
 // }());
 
-
-
-
-
-
 function encrypt(text) {
   const algorithm = 'aes-256-ctr';
   const password = process.env.ENCRYPTION_PASSWORD_00001;
@@ -518,8 +471,6 @@ function decrypt(text) {
   return dec;
 }
 decrypt; // appease linter
-
-
 function makeSessionToken() {
   // These can probably be shortened at some point.
   return crypto.randomBytes(32).toString('base64').replace(/[^A-Za-z0-9]/g, "").substr(0, 20);
@@ -556,8 +507,6 @@ function getUserInfoForSessionToken(sessionToken, res, cb) {
     cb(null, uid);
   });
 }
-
-
 function createPolisLtiToken(tool_consumer_instance_guid, lti_user_id) {
   return ["xPolisLtiToken", tool_consumer_instance_guid, lti_user_id].join(":::");
 }
@@ -585,9 +534,6 @@ function parsePolisLtiToken(token) {
   };
   return o;
 }
-
-
-
 
 function getUserInfoForPolisLtiToken(token) {
   let o = parsePolisLtiToken(token);
@@ -622,8 +568,6 @@ function endSession(sessionToken, cb) {
     cb(null);
   });
 }
-
-
 function setupPwReset(uid, cb) {
   function makePwResetToken() {
     // These can probably be shortened at some point.
@@ -668,19 +612,13 @@ function clearPwResetToken(pwresettoken, cb) {
   });
 }
 
-
-
 function hasAuthToken(req) {
   return !!req.cookies[COOKIES.TOKEN];
 }
 
-
-
 function getUidForApiKey(apikey) {
   return pgQueryP_readOnly_wRetryIfEmpty("select uid from apikeysndvweifu WHERE apikey = ($1);", [apikey]);
 }
-
-
 // http://en.wikipedia.org/wiki/Basic_access_authentication#Client_side
 function doApiKeyBasicAuth(assigner, header, isOptional, req, res, next) {
   let token = header.split(/\s+/).pop() || '', // and the encoded auth token
@@ -712,99 +650,11 @@ function doApiKeyAuth(assigner, apikey, isOptional, req, res, next) {
 //   return pgQueryP("select * from xids where xid = ($2) and owner = (select org_id from conversations where zid = (select zid from zinvites where zinvite = ($1)))", [zinvite, xid]);
 // }
 
-
-function createDummyUser() {
-  return new MPromise("createDummyUser", function(resolve, reject) {
-    pgQuery("INSERT INTO users (created) VALUES (default) RETURNING uid;", [], function(err, results) {
-      if (err || !results || !results.rows || !results.rows.length) {
-        console.error(err);
-        reject(new Error("polis_err_create_empty_user"));
-        return;
-      }
-      resolve(results.rows[0].uid);
-    });
-  });
-}
-
-function createXidRecord(ownerUid, uid, xid, x_profile_image_url, x_name, x_email) {
-  return pgQueryP("insert into xids (owner, uid, xid, x_profile_image_url, x_name, x_email) values ($1, $2, $3, $4, $5, $6) " +
-    "on conflict (owner, xid) do nothing;", [
-      ownerUid,
-      uid,
-      xid,
-      x_profile_image_url || null,
-      x_name || null,
-      x_email || null,
-    ]);
-}
-
-
-
-function getConversationInfo(zid) {
-  return new MPromise("getConversationInfo", function(resolve, reject) {
-    pgQuery("SELECT * FROM conversations WHERE zid = ($1);", [zid], function(err, result) {
-      if (err) {
-        reject(err);
-      } else {
-        resolve(result.rows[0]);
-      }
-    });
-  });
-}
-
-function getConversationInfoByConversationId(conversation_id) {
-  return new MPromise("getConversationInfoByConversationId", function(resolve, reject) {
-    pgQuery("SELECT * FROM conversations WHERE zid = (select zid from zinvites where zinvite = ($1));", [conversation_id], function(err, result) {
-      if (err) {
-        reject(err);
-      } else {
-        resolve(result.rows[0]);
-      }
-    });
-  });
-}
-
-function isXidWhitelisted(owner, xid) {
-  return pgQueryP("select * from xid_whitelist where owner = ($1) and xid = ($2);", [owner, xid]).then((rows) => {
-    return !!rows && rows.length > 0;
-  });
-}
-
-function getXidRecordByXidOwnerId(xid, owner, zid_optional, x_profile_image_url, x_name, x_email, createIfMissing) {
-  return pgQueryP("select * from xids where xid = ($1) and owner = ($2);", [xid, owner]).then(function(rows) {
-    if (!rows || !rows.length) {
-      console.log('no xInfo yet');
-      if (!createIfMissing) {
-        return null;
-      }
-
-      var shouldCreateXidEntryPromise = !zid_optional ? Promise.resolve(true) : getConversationInfo(zid_optional).then((conv) => {
-        return conv.use_xid_whitelist ? isXidWhitelisted(owner, xid) : Promise.resolve(true);
-      });
-
-      return shouldCreateXidEntryPromise.then((should) => {
-        if (!should) {
-          return null;
-        }
-        return createDummyUser().then((newUid) => {
-          console.log('created dummy');
-          return createXidRecord(owner, newUid, xid, x_profile_image_url||null, x_name||null, x_email||null).then(() => {
-            console.log('created xInfo');
-            return [{
-              uid: newUid,
-              owner: owner,
-              xid: xid,
-              x_profile_image_url: x_profile_image_url,
-              x_name: x_name,
-              x_email: x_email,
-            }];
-          });
-        });
-      });
-    }
-    return rows;
-  });
-}
+const createDummyUser = User.createDummyUser;
+const getConversationInfo = Conversation.getConversationInfo;
+const getConversationInfoByConversationId = Conversation.getConversationInfoByConversationId;
+const isXidWhitelisted = Conversation.isXidWhitelisted;
+const getXidRecordByXidOwnerId = Conversation.getXidRecordByXidOwnerId;
 
 // function doXidOwnerConversationIdAuth(assigner, xid, conversation_id, req, res, next) {
 //   getXidRecordByXidConversationId(xid, conversation_id).then(function(rows) {
@@ -817,7 +667,6 @@ function getXidRecordByXidOwnerId(xid, owner, zid_optional, x_profile_image_url,
 //     next();
 //   });
 // }
-
 
 function doXidApiKeyAuth(assigner, apikey, xid, isOptional, req, res, next) {
   getUidForApiKey(apikey).then(function(rows) {
@@ -864,8 +713,6 @@ function doXidApiKeyAuth(assigner, apikey, xid, isOptional, req, res, next) {
     next("polis_err_auth_misc_23423");
   });
 }
-
-
 function doHeaderAuth(assigner, isOptional, req, res, next) {
   let token = req.headers["x-polis"];
 
@@ -912,14 +759,10 @@ function doPolisSlackTeamUserTokenHeaderAuth(assigner, isOptional, req, res, nex
     return;
   });
 }
-
-
 // function logPath(req, res, next) {
 //     winston.log("info",req.method + " " + req.url);
 //     next();
 // }
-
-
 String.prototype.hashCode = function() {
   let hash = 0;
   let i;
@@ -958,11 +801,7 @@ const COOKIES_TO_CLEAR = {
   parent_url: true,
 };
 
-
-
 function initializePolisHelpers() {
-
-
   let polisTypes = {
     reactions: {
       push: 1,
@@ -982,8 +821,6 @@ function initializePolisHelpers() {
   };
   polisTypes.reactionValues = _.values(polisTypes.reactions);
   polisTypes.starValues = _.values(polisTypes.staractions);
-
-
   // // If there are any comments which have no votes by the owner, create a PASS vote by the owner.
   // pgQuery("select * from comments", [], function(err, comments) {
   //     pgQuery("select * from votes", [], function(err, votes) {
@@ -1018,15 +855,11 @@ function initializePolisHelpers() {
   //             function(err, results) {
   //                 winston.log("info",err);
   //             });
-
-
   //         winston.log("info",missing);
   //         winston.log("info",missing.length);
   //         winston.log("info",comments.length);
   //     });
   // });
-
-
 
   let oneYear = 1000 * 60 * 60 * 24 * 365;
 
@@ -1159,55 +992,10 @@ function initializePolisHelpers() {
     });
   }
 
-  // returns a pid of -1 if it's missing
-  function getPid(zid, uid, callback) {
-    let cacheKey = zid + "_" + uid;
-    let cachedPid = pidCache.get(cacheKey);
-    if (!_.isUndefined(cachedPid)) {
-      callback(null, cachedPid);
-      return;
-    }
-    pgQuery_readOnly("SELECT pid FROM participants WHERE zid = ($1) AND uid = ($2);", [zid, uid], function(err, docs) {
-      let pid = -1;
-      if (docs && docs.rows && docs.rows[0]) {
-        pid = docs.rows[0].pid;
-        pidCache.set(cacheKey, pid);
-      }
-      callback(err, pid);
-    });
-  }
-
-
-
-
-  // must follow auth and need('zid'...) middleware
-  function getPidForParticipant(assigner, cache) {
-    return function(req, res, next) {
-      let zid = req.p.zid;
-      let uid = req.p.uid;
-
-      function finish(pid) {
-        assigner(req, "pid", pid);
-        next();
-      }
-      getPidPromise(zid, uid).then(function(pid) {
-        if (pid === -1) {
-          let msg = "polis_err_get_pid_for_participant_missing";
-          yell(msg);
-
-          winston.log("info", zid);
-          winston.log("info", uid);
-          winston.log("info", req.p);
-          next(msg);
-        }
-        finish(pid);
-      }, function(err) {
-        yell("polis_err_get_pid_for_participant");
-        next(err);
-      });
-    };
-  }
-
+  const pidCache = User.pidCache;
+  const getPid = User.getPid;
+  const getPidPromise = User.getPidPromise;
+  const getPidForParticipant = User.getPidForParticipant;
 
   function recordPermanentCookieZidJoin(permanentCookieToken, zid) {
     function doInsert() {
@@ -1237,8 +1025,6 @@ function initializePolisHelpers() {
       language: null,
     }]);
   }
-
-
   if (isTrue(process.env.BACKFILL_COMMENT_LANG_DETECTION)) {
     pgQueryP("select tid, txt, zid from comments where lang is null;", []).then((comments) => {
       let i = 0;
@@ -1264,17 +1050,12 @@ function initializePolisHelpers() {
     });
   }
 
-
-
-
   function translateString(txt, target_lang) {
     if (useTranslateApi) {
       return translateClient.translate(txt, target_lang);
     }
     return Promise.resolve(null);
   }
-
-
   function doVotesPost(uid, pid, conv, tid, voteType, weight, shouldNotify) {
     let zid = conv.zid;
     weight = weight || 0;
@@ -1349,8 +1130,6 @@ function initializePolisHelpers() {
       return doVotesPost(uid, pid, conv, tid, voteType, weight, shouldNotify);
     });
   }
-
-
   function getVotesForSingleParticipant(p) {
     if (_.isUndefined(p.pid)) {
       return Promise.resolve([]);
@@ -1496,8 +1275,6 @@ function initializePolisHelpers() {
       onDone(err);
     });
   }
-
-
   function _auth(assigner, isOptional) {
 
     function getKey(req, key) {
@@ -1591,8 +1368,6 @@ function initializePolisHelpers() {
         }
       });
     }
-
-
     return function(req, res, next) {
 
       doAuth(req, res).then(() => {
@@ -1656,8 +1431,6 @@ function initializePolisHelpers() {
     //   }
     // };
   }
-
-
   // input token from body or query, and populate req.body.u with userid.
   function authOptional(assigner) {
     return _auth(assigner, true);
@@ -1671,8 +1444,6 @@ function initializePolisHelpers() {
     req.body.agid = 1;
     next();
   }
-
-
   /*
   function meter(name) {
       return function (req, res, next){
@@ -1697,16 +1468,12 @@ function initializePolisHelpers() {
       };
   }
   */
-
-
   // 2xx
   // 4xx
   // 5xx
   // logins
   // failed logins
   // forgot password
-
-
 
   let whitelistedCrossDomainRoutes = [
     /^\/api\/v[0-9]+\/launchPrep/,
@@ -1735,16 +1502,12 @@ function initializePolisHelpers() {
     "", // for API
   ];
 
-
-
   let whitelistedBuckets = {
     "pol.is": "pol.is",
     "embed.pol.is": "pol.is",
     "survey.pol.is": "survey.pol.is",
     "preprod.pol.is": "preprod.pol.is",
   };
-
-
   function hasWhitelistMatches(host) {
 
     let hostWithoutProtocol = host;
@@ -1769,8 +1532,6 @@ function initializePolisHelpers() {
     }
     return false;
   }
-
-
   function addCorsHeader(req, res, next) {
 
     let host = "";
@@ -1793,15 +1554,10 @@ function initializePolisHelpers() {
     if (result && result[0]) {
       host = result[0];
     }
-
-
     // check if the route is on a special list that allows it to be called cross domain (by polisHost.js for example)
     let routeIsWhitelistedForAnyDomain = _.some(whitelistedCrossDomainRoutes, function(regex) {
       return regex.test(req.path);
     });
-
-
-
 
     if (!domainOverride && !hasWhitelistMatches(host) && !routeIsWhitelistedForAnyDomain) {
       winston.log("info", 'not whitelisted');
@@ -1820,8 +1576,6 @@ function initializePolisHelpers() {
     }
     return next();
   }
-
-
 
   ////////////////////////////////////////////
   ////////////////////////////////////////////
@@ -1860,11 +1614,7 @@ function initializePolisHelpers() {
     return str;
   }
 
-
-
   function handle_GET_launchPrep(req, res) {
-
-
     let setOnPolisDomain = !domainOverride;
     let origin = req.headers.origin || "";
     if (setOnPolisDomain && origin.match(/^http:\/\/localhost:[0-9]{4}/)) {
@@ -1884,11 +1634,7 @@ function initializePolisHelpers() {
     let dest = hexToStr(req.p.dest);
     res.redirect(dest);
   }
-
-
   // function handle_GET_setFirstCookie(req, res) {
-
-
   //     let setOnPolisDomain = !domainOverride;
   //     let origin = req.headers.origin || "";
   //     if (setOnPolisDomain && origin.match(/^http:\/\/localhost:[0-9]{4}/)) {
@@ -1905,8 +1651,6 @@ function initializePolisHelpers() {
   // });
 
   function handle_GET_tryCookie(req, res) {
-
-
     let setOnPolisDomain = !domainOverride;
     let origin = req.headers.origin || "";
     if (setOnPolisDomain && origin.match(/^http:\/\/localhost:[0-9]{4}/)) {
@@ -1920,8 +1664,6 @@ function initializePolisHelpers() {
     }
     res.status(200).json({});
   }
-
-
   let pcaCacheSize = (process.env.CACHE_MATH_RESULTS === "true") ? 300 : 1;
   let pcaCache = new LruCache({
     max: pcaCacheSize,
@@ -1937,8 +1679,6 @@ function initializePolisHelpers() {
       let timePassed = Date.now() - lastPrefetchPollStartTime;
       return Math.max(0, 2500 - timePassed);
     }
-
-
     // cursor.sort([["math_tick", "asc"]]);
     pgQueryP_readOnly("select * from math_main where caching_tick > ($1) order by caching_tick limit 10;", [lastPrefetchedMathTick]).then((rows) => {
 
@@ -2048,8 +1788,6 @@ function initializePolisHelpers() {
     o['subgroup-clusters'] = _.sortBy(_.map(o['subgroup-clusters'], remapGid), 'id');
     o['subgroup-votes'] = _.sortBy(_.map(o['subgroup-votes'], remapGid), 'id');
     o['subgroup-repness'] = _.sortBy(_.map(o['subgroup-repness'], remapGid), 'id');
-
-
     return o;
   }
   */
@@ -2239,8 +1977,6 @@ function initializePolisHelpers() {
       });
     });
   }
-
-
   function redirectIfHasZidButNoConversationId(req, res, next) {
     if (req.body.zid && !req.body.conversation_id) {
       winston.log("info", "redirecting old zid user to about page");
@@ -2252,8 +1988,6 @@ function initializePolisHelpers() {
     return next();
   }
 
-
-
   function handle_GET_math_pca(req, res) {
     // migrated off this path, old clients were causing timeout issues by polling repeatedly without waiting for a result for a previous poll.
     res.status(304).end();
@@ -2263,8 +1997,6 @@ function initializePolisHelpers() {
   // Needed to determine whether to return a 404 or a 304.
   // zid -> boolean
   let pcaResultsExistForZid = {};
-
-
   function handle_GET_math_pca2(req, res) {
     let zid = req.p.zid;
     let math_tick = req.p.math_tick;
@@ -2285,8 +2017,6 @@ function initializePolisHelpers() {
     } else if (_.isUndefined(math_tick)) {
       math_tick = -1;
     }
-
-
     function finishWith304or404() {
       if (pcaResultsExistForZid[zid]) {
         res.status(304).end();
@@ -2328,8 +2058,6 @@ function initializePolisHelpers() {
       fail(res, 500, err);
     });
   }
-
-
   /*
       addConversationId(o).then(function(item) {
           // ensure we don't expose zid
@@ -2443,11 +2171,7 @@ function initializePolisHelpers() {
     }).catch((err) => {
       return fail(res, 500, "polis_err_GET_math_correlationMatrix", err);
     });
-
-
   }
-
-
 
   function doAddDataExportTask(math_env, email, zid, atDate, format, task_bucket) {
     return pgQueryP("insert into worker_tasks (math_env, task_data, task_type, task_bucket) values ($1, $2, 'generate_export_data', $3);", [
@@ -2461,8 +2185,6 @@ function initializePolisHelpers() {
       task_bucket, // TODO hash the params to get a consistent number?
     ]);
   }
-
-
   if (process.env.RUN_PERIODIC_EXPORT_TESTS && !devMode && process.env.MATH_ENV === "preprod") {
     let runExportTest = () => {
       let math_env = "prod";
@@ -2490,8 +2212,6 @@ function initializePolisHelpers() {
     }
     setInterval(runExportTest, 6 * 60 * 60 * 1000); // every 6 hours
   }
-
-
   function handle_GET_dataExport(req, res) {
 
     getUserInfoForUid2(req.p.uid).then((user) => {
@@ -2512,8 +2232,6 @@ function initializePolisHelpers() {
       fail(res, 500, "polis_err_data_export123b", err);
     });
   }
-
-
   function handle_GET_dataExport_results(req, res) {
 
     var url = s3Client.getSignedUrl('getObject', {
@@ -2528,12 +2246,8 @@ function initializePolisHelpers() {
     // });
     // return res.end();
   }
-
-
   function getBidIndexToPidMapping(zid, math_tick) {
     math_tick = math_tick || -1;
-
-
     return pgQueryP_readOnly("select * from math_bidtopid where zid = ($1) and math_env = ($2);", [zid, process.env.MATH_ENV]).then((rows) => {
 
       if (zid === 12480) {
@@ -2549,8 +2263,6 @@ function initializePolisHelpers() {
       }
     });
   }
-
-
   function handle_GET_bidToPid(req, res) {
     let zid = req.p.zid;
     let math_tick = req.p.math_tick;
@@ -2578,8 +2290,6 @@ function initializePolisHelpers() {
         });
     });
   }
-
-
   function handle_GET_xids(req, res) {
     let uid = req.p.uid;
     let zid = req.p.zid;
@@ -2598,8 +2308,6 @@ function initializePolisHelpers() {
       fail(res, 500, "polis_err_get_xids", err);
     });
   }
-
-
   function handle_POST_xidWhitelist(req, res) {
     const xid_whitelist = req.p.xid_whitelist;
     const len = xid_whitelist.length;
@@ -2619,8 +2327,6 @@ function initializePolisHelpers() {
       return fail(res, 500, "polis_err_POST_xidWhitelist", err);
     });
   }
-
-
   function getBidsForPids(zid, math_tick, pids) {
     let dataPromise = getBidIndexToPidMapping(zid, math_tick);
     let mathResultsPromise = getPca(zid, math_tick);
@@ -2628,8 +2334,6 @@ function initializePolisHelpers() {
     return Promise.all([dataPromise, mathResultsPromise]).then(function(items) {
       let b2p = items[0].bidToPid || []; // not sure yet if "|| []" is right here.
       let mathResults = items[1].asPOJO;
-
-
       function findBidForPid(pid) {
         let yourBidi = -1;
         // if (!b2p) {
@@ -2665,8 +2369,6 @@ function initializePolisHelpers() {
   //     return pcaData.asPOJO["group-clusters"];
   //   });
   // }
-
-
   function handle_GET_bid(req, res) {
     let uid = req.p.uid;
     let zid = req.p.zid;
@@ -2680,8 +2382,6 @@ function initializePolisHelpers() {
       let b2p = items[0].bidToPid || []; // not sure yet if "|| []" is right here.
       let pid = items[1];
       let mathResults = items[2].asPOJO;
-
-
       if (pid < 0) {
         // NOTE: this API should not be called in /demo mode
         fail(res, 500, "polis_err_get_bid_bad_pid");
@@ -2718,8 +2418,6 @@ function initializePolisHelpers() {
     });
   }
 
-
-
   function handle_POST_auth_password(req, res) {
     let pwresettoken = req.p.pwresettoken;
     let newPassword = req.p.newPassword;
@@ -2752,8 +2450,6 @@ function initializePolisHelpers() {
       });
     });
   }
-
-
   function getServerNameWithProtocol(req) {
     let server = "https://pol.is";
 
@@ -2825,8 +2521,6 @@ function initializePolisHelpers() {
       fail(res, 500, "polis_err_slack_oauth", err);
     });
   }
-
-
   function handle_POST_auth_pwresettoken(req, res) {
     let email = req.p.email;
 
@@ -2885,8 +2579,6 @@ Feel free to reply to this email if you need help.`;
     });
   }
 
-
-
   function clearCookie(req, res, cookieName) {
     let origin = req.headers.origin || "";
     if (domainOverride || origin.match(/^http:\/\/localhost:[0-9]{4}/)) {
@@ -2930,8 +2622,6 @@ Feel free to reply to this email if you need help.`;
     }
     winston.log("info", "after clear res set-cookie: " + JSON.stringify(res._headers["set-cookie"]));
   }
-
-
   function doCookieAuth(assigner, isOptional, req, res, next) {
 
     let token = req.cookies[COOKIES.TOKEN];
@@ -2958,8 +2648,6 @@ Feel free to reply to this email if you need help.`;
       next();
     });
   }
-
-
   function handle_POST_auth_deregister(req, res) {
     req.p = req.p || {};
     let token = req.cookies[COOKIES.TOKEN];
@@ -3056,8 +2744,6 @@ Feel free to reply to this email if you need help.`;
       fail(res, 500, "polis_err_metrics_post", err);
     });
   }
-
-
 
   function handle_GET_zinvites(req, res) {
     // if uid is not conversation owner, fail
@@ -3191,8 +2877,6 @@ Feel free to reply to this email if you need help.`;
   //     return pgQueryP("insert into apikeysndvweifu (uid, apikey)  VALUES ($1, $2);", [uid, apikey]);
   //   });
   // }
-
-
   // function getApiKeysTruncated(uid) {
   //   return pgQueryP_readOnly("select * from apikeysndvweifu WHERE uid = ($1);", [uid]).then(function(rows) {
   //     if (!rows || !rows.length) {
@@ -3226,8 +2910,6 @@ Feel free to reply to this email if you need help.`;
   //   apikeyTruncated = apikeyTruncated.replace(/[^a-zA-Z0-9_]/g, 'x');
   //   return pgQueryP("delete from apikeysndvweifu where uid = ($1) and apikey ~ '^" + apikeyTruncated + "';", [uid]);
   // }
-
-
 
   // function addApiKeyForUsersBulk(uids, optionalPrefix) {
   //     let promises = uids.map(function(uid) {
@@ -3271,8 +2953,6 @@ Feel free to reply to this email if you need help.`;
   //     console.error("errorfooooo");
   //     console.error(err);
   // });
-
-
   function generateAndRegisterZinvite(zid, generateShort) {
     let len = 10;
     if (generateShort) {
@@ -3284,8 +2964,6 @@ Feel free to reply to this email if you need help.`;
       });
     });
   }
-
-
 
   function handle_POST_zinvites(req, res) {
     let generateShortUrl = req.p.short_url;
@@ -3634,8 +3312,6 @@ Feel free to reply to this email if you need help.`;
   function updateLastInteractionTimeForConversation(zid, uid) {
     return pgQueryP("update participants set last_interaction = now_as_millis(), nsli = 0 where zid = ($1) and uid = ($2);", [zid, uid]);
   }
-
-
   function populateGeoIpInfo(zid, uid, ipAddress) {
     var userId = process.env.MAXMIND_USERID;
     var licenseKey = process.env.MAXMIND_LICENSEKEY;
@@ -3651,8 +3327,6 @@ Feel free to reply to this email if you need help.`;
       url = "https://geoip.maxmind.com/geoip/v2.1/insights/";
       contentType = "application/vnd.maxmind.com-insights+json; charset=UTF-8; version=2.1";
     }
-
-
     return request.get(url + ipAddress, {
       method: "GET",
       contentType: contentType,
@@ -3678,8 +3352,6 @@ Feel free to reply to this email if you need help.`;
 
     });
   }
-
-
 
   function addExtendedParticipantInfo(zid, uid, data) {
     if (!data || !_.keys(data).length) {
@@ -3862,8 +3534,6 @@ Feel free to reply to this email if you need help.`;
       });
     });
   }
-
-
   function getAnswersForConversation(zid, callback) {
     pgQuery_readOnly("SELECT * from participant_metadata_answers WHERE zid = ($1) AND alive=TRUE;", [zid], function(err, x) {
       if (err) {
@@ -3890,35 +3560,8 @@ Feel free to reply to this email if you need help.`;
     });
   }
 
-
-  function getUserInfoForUid(uid, callback) {
-    pgQuery_readOnly("SELECT email, hname from users where uid = $1", [uid], function(err, results) {
-      if (err) {
-        return callback(err);
-      }
-      if (!results.rows || !results.rows.length) {
-        return callback(null);
-      }
-      callback(null, results.rows[0]);
-    });
-  }
-
-  function getUserInfoForUid2(uid) {
-    return new MPromise("getUserInfoForUid2", function(resolve, reject) {
-      pgQuery_readOnly("SELECT * from users where uid = $1", [uid], function(err, results) {
-        if (err) {
-          return reject(err);
-        }
-        if (!results.rows || !results.rows.length) {
-          return reject(null);
-        }
-        let o = results.rows[0];
-        resolve(o);
-      });
-    });
-  }
-
-
+  const getUserInfoForUid = User.getUserInfoForUid;
+  const getUserInfoForUid2 = User.getUserInfoForUid2;
 
   function emailFeatureRequest(message) {
     const body =
@@ -3954,8 +3597,6 @@ ${message}`;
 
     return emailTeam("Polis Bad Problems!!!", body);
   }
-
-
   function sendPasswordResetEmail(uid, pwresettoken, serverName, callback) {
     getUserInfoForUid(uid, function(err, userInfo) {
       if (err) {
@@ -3988,9 +3629,6 @@ ${serverName}/pwreset/${pwresettoken}
     });
   }
 
-
-
-
   // function sendTextEmailWithPostmark(sender, recipient, subject, text) {
   //   winston.log("info", "sending email with postmark: " + [sender, recipient, subject, text].join(" "));
   //   return new Promise(function(resolve, reject) {
@@ -4011,8 +3649,6 @@ ${serverName}/pwreset/${pwresettoken}
   //     });
   //   });
   // }
-
-
   function sendMultipleTextEmails(sender, recipientArray, subject, text) {
     recipientArray = recipientArray || [];
     return Promise.all(recipientArray.map(function(email) {
@@ -4041,8 +3677,6 @@ ${serverName}/pwreset/${pwresettoken}
   }
   setInterval(trySendingBackupEmailTest, 1000 * 60 * 60 * 23); // try every 23 hours (so it should only try roughly once a day)
   trySendingBackupEmailTest();
-
-
   function sendEinviteEmail(req, email, einvite) {
     let serverName = getServerNameWithProtocol(req);
     const body =
@@ -4078,7 +3712,7 @@ ${serverName}/api/v3/verify?e=${einvite}`;
   }
 
   function isEmailVerified(email) {
-    return pgQueryP("select * from email_validations where email = ($1);", [email]).then(function(rows) {
+    return pg.pgQueryP("select * from email_validations where email = ($1);", [email]).then(function(rows) {
       return rows.length > 0;
     });
   }
@@ -4165,8 +3799,6 @@ Email verified! You can close this tab or hit the back button.
     });
   }
 
-
-
   function handle_GET_participants(req, res) {
     // let pid = req.p.pid;
     let uid = req.p.uid;
@@ -4211,15 +3843,11 @@ Email verified! You can close this tab or hit the back button.
 
     // });
   }
-
-
   function handle_GET_dummyButton(req, res) {
     let message = req.p.button + " " + req.p.uid;
     emailFeatureRequest(message);
     res.status(200).end();
   }
-
-
   function doGetConversationsRecent(req, res, field) {
     if (!isPolisDev(req.p.uid)) {
       fail(res, 403, "polis_err_no_access_for_this_user");
@@ -4247,8 +3875,6 @@ Email verified! You can close this tab or hit the back button.
     doGetConversationsRecent(req, res, "modified");
   }
 
-
-
   function userHasAnsweredZeQuestions(zid, answers) {
     return new MPromise("userHasAnsweredZeQuestions", function(resolve, reject) {
       getAnswersForConversation(zid, function(err, available_answers) {
@@ -4273,8 +3899,6 @@ Email verified! You can close this tab or hit the back button.
       });
     });
   }
-
-
   function handle_POST_participants(req, res) {
     let zid = req.p.zid;
     let uid = req.p.uid;
@@ -4308,8 +3932,6 @@ Email verified! You can close this tab or hit the back button.
     }
 
     function doJoin() {
-
-
 
       userHasAnsweredZeQuestions(zid, answers).then(function() {
         joinConversation(zid, uid, info, answers).then(function(ptpt) {
@@ -4363,24 +3985,6 @@ Email verified! You can close this tab or hit the back button.
     });
   }
 
-  function addLtiUserifNeeded(uid, lti_user_id, tool_consumer_instance_guid, lti_user_image) {
-    lti_user_image = lti_user_image || null;
-    return pgQueryP("select * from lti_users where lti_user_id = ($1) and tool_consumer_instance_guid = ($2);", [lti_user_id, tool_consumer_instance_guid]).then(function(rows) {
-      if (!rows || !rows.length) {
-        return pgQueryP("insert into lti_users (uid, lti_user_id, tool_consumer_instance_guid, lti_user_image) values ($1, $2, $3, $4);", [uid, lti_user_id, tool_consumer_instance_guid, lti_user_image]);
-      }
-    });
-  }
-
-  function addLtiContextMembership(uid, lti_context_id, tool_consumer_instance_guid) {
-    return pgQueryP("select * from lti_context_memberships where uid = $1 and lti_context_id = $2 and tool_consumer_instance_guid = $3;", [uid, lti_context_id, tool_consumer_instance_guid]).then(function(rows) {
-      if (!rows || !rows.length) {
-        return pgQueryP("insert into lti_context_memberships (uid, lti_context_id, tool_consumer_instance_guid) values ($1, $2, $3);", [uid, lti_context_id, tool_consumer_instance_guid]);
-      }
-    });
-  }
-
-
   function checkPassword(uid, password) {
     return pgQueryP_readOnly_wRetryIfEmpty("select pwhash from jianiuevyew where uid = ($1);", [uid]).then(function(rows) {
       if (!rows || !rows.length) {
@@ -4400,6 +4004,9 @@ Email verified! You can close this tab or hit the back button.
       });
     });
   }
+
+  const addLtiUserifNeeded = User.addLtiUserIfNeeded;
+  const addLtiContextMembership = User.addLtiContextMembership;
 
   function subscribeToNotifications(zid, uid, email) {
     let type = 1; // 1 for email
@@ -4551,8 +4158,6 @@ Email verified! You can close this tab or hit the back button.
       return shouldTryAgain;
     });
   }
-
-
   function doNotificationBatch() {
     return claimNextNotificationTask().then((task) => {
       if (!task) {
@@ -4599,8 +4204,6 @@ Email verified! You can close this tab or hit the back button.
   if (shouldSendNotifications) {
     doNotificationLoop();
   }
-
-
   function createNotificationsUnsubscribeUrl(conversation_id, email) {
     let params = {
       conversation_id: conversation_id,
@@ -4631,8 +4234,6 @@ Email verified! You can close this tab or hit the back button.
     return server + "/" + path + "?" + paramsToStringSortedByName(params);
   }
 
-
-
   function handle_GET_notifications_subscribe(req, res) {
     let zid = req.p.zid;
     let email = req.p.email;
@@ -4657,8 +4258,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 500, "polis_err_subscribe_misc", err);
     });
   }
-
-
   function handle_GET_notifications_unsubscribe(req, res) {
     let zid = req.p.zid;
     let email = req.p.email;
@@ -4683,8 +4282,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 500, "polis_err_unsubscribe_misc", err);
     });
   }
-
-
   function handle_POST_convSubscriptions(req, res) {
     let zid = req.p.zid;
     let uid = req.p.uid;
@@ -4710,8 +4307,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 400, "polis_err_bad_subscription_type", new Error("polis_err_bad_subscription_type"));
     }
   }
-
-
 
   function handle_POST_auth_login(req, res) {
     let password = req.p.password;
@@ -4776,7 +4371,7 @@ Email verified! You can close this tab or hit the back button.
               winston.log("info", "lti_user_id", lti_user_id);
               winston.log("info", "lti_context_id", lti_context_id);
               let ltiUserPromise = lti_user_id ?
-                addLtiUserifNeeded(uid, lti_user_id, tool_consumer_instance_guid, lti_user_image) :
+                addLtiUserIfNeeded(uid, lti_user_id, tool_consumer_instance_guid, lti_user_image) :
                 Promise.resolve();
               let ltiContextMembershipPromise = lti_context_id ?
                 addLtiContextMembership(uid, lti_context_id, tool_consumer_instance_guid) :
@@ -4786,7 +4381,7 @@ Email verified! You can close this tab or hit the back button.
                   if (afterJoinRedirectUrl) {
                     res.redirect(afterJoinRedirectUrl);
                   } else {
-                    renderLtiLinkageSuccessPage(req, res, {
+                    User.renderLtiLinkageSuccessPage(req, res, {
                       // may include token here too
                       context_id: lti_context_id,
                       uid: uid,
@@ -4810,8 +4405,6 @@ Email verified! You can close this tab or hit the back button.
     }); // users query
   } // /api/v3/auth/login
 
-
-
   function handle_POST_joinWithInvite(req, res) {
 
     // if they're already in the conv
@@ -4826,8 +4419,6 @@ Email verified! You can close this tab or hit the back button.
     //              force them to create a full account
     //     else
     //         let them join without forcing a sign in (assuming conversation allows that)
-
-
 
     return joinWithZidOrSuzinvite({
       answers: req.p.answers,
@@ -4881,8 +4472,6 @@ Email verified! You can close this tab or hit the back button.
       }
     });
   }
-
-
   // Test for deadlock condition
   // _.times(2, function() {
   // setInterval(function() {
@@ -4901,8 +4490,6 @@ Email verified! You can close this tab or hit the back button.
 
   // }, 10);
   // });
-
-
   function joinWithZidOrSuzinvite(o) {
     return Promise.resolve(o)
       .then(function(o) {
@@ -5032,8 +4619,6 @@ Email verified! You can close this tab or hit the back button.
         }
       });
   }
-
-
   function startSessionAndAddCookies(req, res, uid) {
     return new Promise(function(resolve, reject) {
       startSession(uid, function(err, token) {
@@ -5044,36 +4629,6 @@ Email verified! You can close this tab or hit the back button.
         resolve(addCookies(req, res, token, uid));
       });
     });
-  }
-
-  function renderLtiLinkageSuccessPage(req, res, o) {
-    res.set({
-      'Content-Type': 'text/html',
-    });
-    let html = "" +
-      "<!DOCTYPE html><html lang='en'>" +
-      '<head>' +
-      '<meta name="viewport" content="width=device-width, initial-scale=1;">' +
-      '</head>' +
-      "<body style='max-width:320px'>" +
-      "<p>You are signed in as polis user " + o.email + "</p>" +
-      // "<p><a href='https://pol.is/user/logout'>Change pol.is users</a></p>" +
-      // "<p><a href='https://preprod.pol.is/inbox/context="+ o.context_id +"'>inbox</a></p>" +
-      // "<p><a href='https://preprod.pol.is/2demo' target='_blank'>2demo</a></p>" +
-      // "<p><a href='https://preprod.pol.is/conversation/create/context="+ o.context_id +"'>create</a></p>" +
-
-      // form for sign out
-      '<p><form role="form" class="FormVertical" action="' + getServerNameWithProtocol(req) + '/api/v3/auth/deregister" method="POST">' +
-      '<input type="hidden" name="showPage" value="canvas_assignment_deregister">' +
-      '<button type="submit" class="Btn Btn-primary">Change pol.is users</button>' +
-      '</form></p>' +
-
-      // "<p style='background-color: yellow;'>" +
-      //     JSON.stringify(req.body)+
-      //     (o.user_image ? "<img src='"+o.user_image+"'></img>" : "") +
-      // "</p>"+
-      "</body></html>";
-    res.status(200).send(html);
   }
 
   function deleteFacebookUserRecord(o) {
@@ -5156,8 +4711,6 @@ Email verified! You can close this tab or hit the back button.
     }
   }
 
-
-
   function handle_GET_perfStats(req, res) {
     res.json(METRICS_IN_RAM);
   }
@@ -5175,8 +4728,6 @@ Email verified! You can close this tab or hit the back button.
     }
     return firstVotes;
   }
-
-
   function isParentDomainWhitelisted(domain, zid, isWithinIframe, domain_whitelist_override_key) {
     return pgQueryP_readOnly(
         "select * from site_domain_whitelist where site_id = " +
@@ -5256,13 +4807,9 @@ Email verified! You can close this tab or hit the back button.
         return ok;
       });
   }
-
-
   function denyIfNotFromWhitelistedDomain(req, res, next) {
 
     let isWithinIframe = req.headers && req.headers.referer && req.headers.referer.includes('parent_url');
-
-
     // res.status(403);
     // next("polis_err_domain");
     // return;
@@ -5281,14 +4828,10 @@ Email verified! You can close this tab or hit the back button.
       ref = ref && ref.length >= 3 && ref[2];
       ref = ref || "";
     }
-
-
     // let path = req.path;
     // path = path && path.split('/');
     // let conversation_id = path && path.length >= 2 && path[1];
     let zid = req.p.zid;
-
-
 
     isParentDomainWhitelisted(ref, zid, isWithinIframe, req.p.domain_whitelist_override_key).then(function(isOk) {
       if (isOk) {
@@ -5303,8 +4846,6 @@ Email verified! You can close this tab or hit the back button.
       next("polis_err_domain_misc");
     });
   }
-
-
 
   function setDomainWhitelist(uid, newWhitelist) {
     // TODO_UPSERT
@@ -5326,8 +4867,6 @@ Email verified! You can close this tab or hit the back button.
       return rows[0].domain_whitelist;
     });
   }
-
-
   function handle_GET_domainWhitelist(req, res) {
     getDomainWhitelist(req.p.uid).then(function(whitelist) {
       res.json({
@@ -5337,8 +4876,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 500, "polis_err_get_domainWhitelist_misc", err);
     });
   }
-
-
   function handle_POST_domainWhitelist(req, res) {
     setDomainWhitelist(req.p.uid, req.p.domain_whitelist).then(function() {
       res.json({
@@ -5348,8 +4885,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 500, "polis_err_post_domainWhitelist_misc", err);
     });
   }
-
-
   function handle_GET_conversationStats(req, res) {
     let zid = req.p.zid;
     let uid = req.p.uid;
@@ -5477,8 +5012,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 500, "polis_err_conversationStats_misc", err);
     });
   }
-
-
   function handle_GET_snapshot(req, res) {
     let uid = req.p.uid;
     let zid = req.p.zid;
@@ -5486,8 +5019,6 @@ Email verified! You can close this tab or hit the back button.
     if (true) {
       throw new Error("TODO Needs to clone participants_extended and any other new tables as well.");
     }
-
-
     if (isPolisDev(uid)) {
       // is polis developer
     } else {
@@ -5544,8 +5075,6 @@ Email verified! You can close this tab or hit the back button.
           });
       });
   }
-
-
   function handle_GET_facebook_delete(req, res) {
     deleteFacebookUserRecord(req.p).then(function() {
       res.json({});
@@ -5553,8 +5082,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 500, err);
     });
   }
-
-
   function getFriends(fb_access_token) {
     function getMoreFriends(friendsSoFar, urlForNextCall) {
       // urlForNextCall includes access token
@@ -5607,8 +5134,6 @@ Email verified! You can close this tab or hit the back button.
       }
     });
   }
-
-
   function handle_POST_auth_facebook(req, res) {
     let response = JSON.parse(req.p.response);
     let fb_access_token = response && response.authResponse && response.authResponse.accessToken;
@@ -5671,8 +5196,6 @@ Email verified! You can close this tab or hit the back button.
       });
     });
   }
-
-
   function do_handle_POST_auth_facebook(req, res, o) {
 
     // If a pol.is user record exists, and someone logs in with a facebook account that has the same email address, we should bind that facebook account to the pol.is account, and let the user sign in.
@@ -5707,8 +5230,6 @@ Email verified! You can close this tab or hit the back button.
       fb_friends_response: req.p.fb_friends_response || "",
       response: req.p.response,
     };
-
-
     function doFbUserHasAccountLinked(user) {
       if (user.fb_user_id === fb_user_id) {
         updateFacebookUserRecord(Object.assign({}, {
@@ -5932,8 +5453,6 @@ Email verified! You can close this tab or hit the back button.
         });
     });
   } // end do_handle_POST_auth_facebook
-
-
   function handle_POST_auth_new(req, res) {
     let hname = req.p.hname;
     let password = req.p.password;
@@ -6026,8 +5545,6 @@ Email verified! You can close this tab or hit the back button.
               fail(res, 500, "polis_err_reg_failed_to_add_user_record", err);
               return;
             }
-
-
             startSession(uid, function(err, token) {
               if (err) {
                 fail(res, 500, "polis_err_reg_failed_to_start_session", err);
@@ -6036,7 +5553,7 @@ Email verified! You can close this tab or hit the back button.
               addCookies(req, res, token, uid).then(function() {
 
                 let ltiUserPromise = lti_user_id ?
-                  addLtiUserifNeeded(uid, lti_user_id, tool_consumer_instance_guid, lti_user_image) :
+                  addLtiUserIfNeeded(uid, lti_user_id, tool_consumer_instance_guid, lti_user_image) :
                   Promise.resolve();
                 let ltiContextMembershipPromise = lti_context_id ?
                   addLtiContextMembership(uid, lti_context_id, tool_consumer_instance_guid) :
@@ -6046,7 +5563,7 @@ Email verified! You can close this tab or hit the back button.
                     if (afterJoinRedirectUrl) {
                       res.redirect(afterJoinRedirectUrl);
                     } else {
-                      renderLtiLinkageSuccessPage(req, res, {
+                      User.renderLtiLinkageSuccessPage(req, res, {
                         // may include token here too
                         context_id: lti_context_id,
                         uid: uid,
@@ -6079,8 +5596,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 500, "polis_err_reg_checking_existing_users", err);
     });
   } // end /api/v3/auth/new
-
-
   function handle_POST_tutorial(req, res) {
     let uid = req.p.uid;
     let step = req.p.step;
@@ -6090,8 +5605,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 500, "polis_err_saving_tutorial_state", err);
     });
   }
-
-
 
   function handle_GET_users(req, res) {
     let uid = req.p.uid;
@@ -6110,66 +5623,7 @@ Email verified! You can close this tab or hit the back button.
     });
   }
 
-  function getUser(uid, zid_optional, xid_optional, owner_uid_optional) {
-    if (!uid) {
-      // this api may be called by a new user, so we don't want to trigger a failure here.
-      return Promise.resolve({});
-    }
-
-    let xidInfoPromise = Promise.resolve(null);
-    if (zid_optional && xid_optional) {
-      xidInfoPromise = getXidRecord(xid_optional, zid_optional);
-    } else if (xid_optional && owner_uid_optional) {
-      xidInfoPromise = getXidRecordByXidOwnerId(xid_optional, owner_uid_optional, zid_optional);
-    }
-
-    return Promise.all([
-      getUserInfoForUid2(uid),
-      getFacebookInfo([uid]),
-      getTwitterInfo([uid]),
-      xidInfoPromise,
-    ]).then(function(o) {
-      let info = o[0];
-      let fbInfo = o[1];
-      let twInfo = o[2];
-      let xInfo = o[3];
-
-      let hasFacebook = fbInfo && fbInfo.length && fbInfo[0];
-      let hasTwitter = twInfo && twInfo.length && twInfo[0];
-      let hasXid = xInfo && xInfo.length && xInfo[0];
-      if (hasFacebook) {
-        let width = 40;
-        let height = 40;
-        fbInfo.fb_picture = "https://graph.facebook.com/v2.2/" + fbInfo.fb_user_id + "/picture?width=" + width + "&height=" + height;
-        delete fbInfo[0].response;
-      }
-      if (hasTwitter) {
-        delete twInfo[0].response;
-      }
-      if (hasXid) {
-        delete xInfo[0].owner;
-        delete xInfo[0].created;
-        delete xInfo[0].uid;
-      }
-      return {
-        uid: uid,
-        email: info.email,
-        hname: info.hname,
-        hasFacebook: !!hasFacebook,
-        facebook: fbInfo && fbInfo[0],
-        twitter: twInfo && twInfo[0],
-        hasTwitter: !!hasTwitter,
-        hasXid: !!hasXid,
-        xInfo: xInfo && xInfo[0],
-        finishedTutorial: !!info.tut,
-        site_ids: [info.site_id],
-        created: Number(info.created),
-        daysInTrial: 10 + (usersToAdditionalTrialDays[uid] || 0),
-        // plan: planCodeToPlanName[info.plan],
-        planCode: info.plan,
-      };
-    });
-  }
+  const getUser = User.getUser;
 
   // These map from non-ui string codes to number codes used in the DB
   // The string representation ("sites", etc) is also used in intercom.
@@ -6196,8 +5650,6 @@ Email verified! You can close this tab or hit the back button.
   //   100: "Site",
   //   1000: "Organization",
   // };
-
-
   function changePlan(uid, planCode) {
     return new Promise(function(resolve, reject) {
       pgQuery("update users set plan = ($1) where uid = ($2);", [planCode, uid], function(err, results) {
@@ -6227,8 +5679,6 @@ Email verified! You can close this tab or hit the back button.
       });
     });
   }
-
-
   function createStripeUser(o) {
     return new Promise(function(resolve, reject) {
       stripe.customers.create(o, function(err, customer) {
@@ -6291,8 +5741,6 @@ Email verified! You can close this tab or hit the back button.
     });
   }
 
-
-
   function handle_GET_createPlanChangeCoupon(req, res) {
     var uid = req.p.uid;
     var planCode = req.p.planCode;
@@ -6308,8 +5756,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 500, "polis_err_creating_coupon_code", err);
     });
   }
-
-
   function handle_GET_changePlanWithCoupon(req, res) {
     var uid = req.p.uid;
     var code = req.p.code;
@@ -6352,8 +5798,6 @@ Email verified! You can close this tab or hit the back button.
       setPlanCookie(req, res, setOnPolisDomain, planCode);
     });
   }
-
-
   function updatePlanOld(req, res, uid, planCode, isCurrentUser) {
     winston.log("info", 'updatePlan', uid, planCode);
     setUsersPlanInIntercom(uid, planCode).catch(function(err) {
@@ -6392,15 +5836,11 @@ Email verified! You can close this tab or hit the back button.
     });
   }
 
-
-
   function handle_POST_stripe_save_token(req, res) {
     console.log("info", "XXX - Got the params!");
     console.log(req.body);
     console.log("info", "XXX - Got the params!");
   }
-
-
   function handle_POST_stripe_upgrade(req, res) {
     var uid = req.p.uid;
     var planName = req.p.plan;
@@ -6415,8 +5855,6 @@ Email verified! You can close this tab or hit the back button.
         "planCode " + planCode + "\n";
 
       emailTeam("Polis account upgraded", body);
-
-
       return updateStripePlan(user, stripeResponse.id, user.email, planName);
     }).then(function() {
       return updatePlan(req, res, uid, planCode);
@@ -6440,8 +5878,6 @@ Email verified! You can close this tab or hit the back button.
 
   function handle_POST_stripe_cancel(req, res) {
     const uid = req.p.uid;
-
-
     getUserInfoForUid2(uid).then((user) => {
 
       emailBadProblemTime("User cancelled subscription: " + user.email);
@@ -6662,8 +6098,6 @@ Email verified! You can close this tab or hit the back button.
       "total as (select count(*) as total from c) " +
       "select cast(remaining.remaining as integer), cast(total.total as integer), cast(($2) as integer) as pid from remaining, total;", [zid, pid]);
   }
-
-
   function getComments(o) {
     let commentListPromise = o.moderation ? _getCommentsForModerationList(o) : _getCommentsList(o);
     let convPromise = getConversationInfo(o.zid);
@@ -6785,8 +6219,6 @@ Email verified! You can close this tab or hit the back button.
 
   */
 
-
-
   function handle_GET_participation(req, res) {
     let zid = req.p.zid;
     let uid = req.p.uid;
@@ -6861,8 +6293,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 500, "polis_err_get_participation_misc", err);
     });
   }
-
-
   function getAgeRange(demo) {
     var currentYear = (new Date()).getUTCFullYear();
     var birthYear = demo.ms_birth_year_estimate_fb;
@@ -6898,9 +6328,6 @@ Email verified! You can close this tab or hit the back button.
     return gender;
   }
 
-
-
-
   function getDemographicsForVotersOnComments(zid, comments) {
     function isAgree(v) {
       return v.vote === polisTypes.reactions.pull;
@@ -6929,8 +6356,6 @@ Email verified! You can close this tab or hit the back button.
     ]).then((a) => {
       var votes = a[0];
       var demo = a[1];
-
-
       demo = demo.map((d) => {
         return {
           pid: d.pid,
@@ -7010,8 +6435,6 @@ Email verified! You can close this tab or hit the back button.
     }
     return Promise.resolve(null);
   }
-
-
   function handle_GET_comments_translations(req, res) {
     const zid = req.p.zid;
     const tid = req.p.tid;
@@ -7090,8 +6513,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 500, "polis_err_get_comments", err);
     });
   } // end GET /api/v3/comments
-
-
   function isDuplicateKey(err) {
     let isdup = err.code === 23505 ||
       err.code === '23505' ||
@@ -7217,8 +6638,6 @@ Email verified! You can close this tab or hit the back button.
       return (rows && rows[0]) || null;
     });
   }
-
-
   // function muteComment(zid, tid) {
   //     let mod = polisTypes.mod.ban;
   //     return moderateComment(zid, tid, false, mod);
@@ -7285,8 +6704,6 @@ Email verified! You can close this tab or hit the back button.
   //         fail(res, 500, err);
   //     });
   // }
-
-
   function hasBadWords(txt) {
     txt = txt.toLowerCase();
     let tokens = txt.split(" ");
@@ -7298,8 +6715,6 @@ Email verified! You can close this tab or hit the back button.
     return false;
   }
 
-
-
   function commentExists(zid, txt) {
     return pgQueryP("select zid from comments where zid = ($1) and txt = ($2);", [zid, txt]).then(function(rows) {
       return rows && rows.length;
@@ -7309,8 +6724,6 @@ Email verified! You can close this tab or hit the back button.
   function handle_POST_comments_slack(req, res) {
     const slack_team = req.p.slack_team;
     const slack_user_id = req.p.slack_user_id;
-
-
     pgQueryP("select * from slack_users where slack_team = ($1) and slack_user_id = ($2);", [slack_team, slack_user_id]).then((rows) => {
       if (!rows || !rows.length) {
         const uidPromise = createDummyUser();
@@ -7338,14 +6751,10 @@ Email verified! You can close this tab or hit the back button.
       req.p.uid = uid;
 
       handle_POST_comments(req, res);
-
-
     }).catch((err) => {
       fail(res, 500, "polis_err_post_comments_slack_misc", err);
     });
   }
-
-
 
   function handle_POST_comments(req, res) {
     let zid = req.p.zid;
@@ -7407,8 +6816,6 @@ Email verified! You can close this tab or hit the back button.
       console.log("POST_comments doGetPid done", Date.now());
       return Promise.resolve(pid);
     }
-
-
     let twitterPrepPromise = Promise.resolve();
     if (twitter_tweet_id) {
       twitterPrepPromise = prepForTwitterComment(twitter_tweet_id, zid);
@@ -7459,13 +6866,7 @@ Email verified! You can close this tab or hit the back button.
 
       let conversationInfoPromise = getConversationInfo(zid);
 
-
-
-
     // return xidUserPromise.then(function(xidUser) {
-
-
-
 
       let shouldCreateXidRecord = false;
 
@@ -7714,8 +7115,6 @@ Email verified! You can close this tab or hit the back button.
     ////}); // nextTick
     //}); // BEGIN
   } // end POST /api/v3/comments
-
-
   function handle_GET_votes_me(req, res) {
     getPid(req.p.zid, req.p.uid, function(err, pid) {
       if (err || pid < 0) {
@@ -7770,8 +7169,6 @@ Email verified! You can close this tab or hit the back button.
       //}
     //});
   //}
-
-
   function selectProbabilistically(comments, priorities, nTotal, nRemaining) {
     // Here we go through all of the comments we might select for the user and add their priority values
     let lookup = _.reduce(comments, (o, comment) => {
@@ -7825,8 +7222,6 @@ Email verified! You can close this tab or hit the back button.
       return c;
     });
   }
-
-
   // function getNextCommentPrioritizingNonPassedComments(zid, pid, withoutTids) {
   //   if (!withoutTids || !withoutTids.length) {
   //     withoutTids = [-999999]; // ensure there's a value in there so the sql parses as a list
@@ -7939,8 +7334,6 @@ Email verified! You can close this tab or hit the back button.
       "(select count(*) from votes where zid = ($1) and pid = ($2)))", [zid, pid]);
   }
 
-
-
   function handle_GET_nextComment(req, res) {
     if (req.timedout) {
       return;
@@ -7980,8 +7373,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 500, "polis_err_get_next_comment", err);
     });
   }
-
-
   function handle_GET_participationInit(req, res) {
 
     // let qs = {
@@ -8111,8 +7502,6 @@ Email verified! You can close this tab or hit the back button.
     });
   }
 
-
-
   function updateConversationModifiedTime(zid, t) {
     let modified = _.isUndefined(t) ? Date.now() : Number(t);
     let query = "update conversations set modified = ($2) where zid = ($1) and modified < ($2);";
@@ -8124,47 +7513,8 @@ Email verified! You can close this tab or hit the back button.
     return pgQueryP(query, params);
   }
 
-
-
-  function createXidRecordByZid(zid, uid, xid, x_profile_image_url, x_name, x_email) {
-    return getConversationInfo(zid).then((conv) => {
-      const shouldCreateXidRecord = conv.use_xid_whitelist ? isXidWhitelisted(conv.owner, xid) : Promise.resolve(true);
-      return shouldCreateXidRecord.then((should) => {
-        if (!should) {
-          throw new Error("polis_err_xid_not_whitelisted_2");
-        }
-        return pgQueryP("insert into xids (owner, uid, xid, x_profile_image_url, x_name, x_email) values ((select org_id from conversations where zid = ($1)), $2, $3, $4, $5, $6) " +
-          "on conflict (owner, xid) do nothing;", [
-            zid,
-            uid,
-            xid,
-            x_profile_image_url || null,
-            x_name || null,
-            x_email || null,
-          ]);
-      });
-    });
-  }
-
-  function getXidRecord(xid, zid) {
-    return pgQueryP("select * from xids where xid = ($1) and owner = (select org_id from conversations where zid = ($2));", [xid, zid]);
-  }
-
-  function getXidStuff(xid, zid) {
-    return getXidRecord(xid, zid).then((rows) => {
-      if (!rows || !rows.length) {
-        return "noXidRecord";
-      }
-      let xidRecordForPtpt = rows[0];
-      if (xidRecordForPtpt) {
-        return getPidPromise(zid, xidRecordForPtpt.uid, true).then((pidForXid) => {
-          xidRecordForPtpt.pid = pidForXid;
-          return xidRecordForPtpt;
-        });
-      }
-      return xidRecordForPtpt;
-    });
-  }
+  const createXidRecordByZid = Conversation.createXidRecordByZid;
+  const getXidStuff = Conversation.getXidStuff;
 
   function handle_PUT_participants_extended(req, res) {
     let zid = req.p.zid;
@@ -8213,8 +7563,6 @@ Email verified! You can close this tab or hit the back button.
       let ptpt = rows[0];
       pid = ptpt.pid;
     }) : Promise.resolve();
-
-
     pidReadyPromise.then(function() {
 
       // let conv;
@@ -8256,8 +7604,6 @@ Email verified! You can close this tab or hit the back button.
         }
         // PID_FLOW This may be the first time the client gets the pid.
         result.currentPid = pid;
-
-
         // result.shouldMod = true; // TODO
         if (result.shouldMod) {
           result.modOptions = {};
@@ -8292,8 +7638,6 @@ Email verified! You can close this tab or hit the back button.
     });
   }
 
-
-
   function handle_POST_ptptCommentMod(req, res) {
     let zid = req.p.zid;
     let pid = req.p.pid;
@@ -8303,8 +7647,6 @@ Email verified! You can close this tab or hit the back button.
     // need('as_important', getBool, assignToP, false),
     // need('as_spam', getBool, assignToP, false),
     // need('as_offtopic', getBool, assignToP, false),
-
-
 
     return pgQueryP("insert into crowd_mod (" +
       "zid, " +
@@ -8376,8 +7718,6 @@ Email verified! You can close this tab or hit the back button.
       });
   }
 
-
-
   function handle_POST_upvotes(req, res) {
     let uid = req.p.uid;
     let zid = req.p.zid;
@@ -8400,8 +7740,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 500, "polis_err_upvote_check", err);
     });
   }
-
-
   function addStar(zid, tid, pid, starred, created) {
     starred = starred ? 1 : 0;
     let query = "INSERT INTO stars (pid, zid, tid, starred, created) VALUES ($1, $2, $3, $4, default) RETURNING created;";
@@ -8412,8 +7750,6 @@ Email verified! You can close this tab or hit the back button.
     }
     return pgQueryP(query, params);
   }
-
-
   function handle_POST_stars(req, res) {
     addStar(req.p.zid, req.p.tid, req.p.pid, req.p.starred).then(function(result) {
       let createdTime = result.rows[0].created;
@@ -8453,8 +7789,6 @@ Email verified! You can close this tab or hit the back button.
       res.status(200).json({}); // TODO don't stop after the first one, map the inserts to deferreds.
     });
   }
-
-
   function verifyMetadataAnswersExistForEachQuestion(zid) {
     let errorcode = "polis_err_missing_metadata_answers";
     return new Promise(function(resolve, reject) {
@@ -8544,8 +7878,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 500, "polis_err_POST_reportCommentSelections_misc", err);
     });
   }
-
-
   // kind of crappy that we're replacing the zinvite.
   // This is needed because we initially create a conversation with the POST, then actually set the properties with the subsequent PUT.
   // if we stop doing that, we can remove this function.
@@ -8711,8 +8043,6 @@ Email verified! You can close this tab or hit the back button.
       res.status(200).json("INSERT INTO lti_oauthv1_credentials (uid, oauth_consumer_key, oauth_shared_secret) values (" + x + ") returning oauth_consumer_key, oauth_shared_secret;");
     });
   }
-
-
   function handle_POST_conversation_close(req, res) {
     var q = "select * from conversations where zid = ($1)";
     var params = [req.p.zid];
@@ -8900,8 +8230,6 @@ Email verified! You can close this tab or hit the back button.
       ifDefinedSet("auth_opt_fb", req.p, fields);
       ifDefinedSet("auth_opt_tw", req.p, fields);
       ifDefinedSet("auth_opt_allow_3rdparty", req.p, fields);
-
-
 
       if (!_.isUndefined(req.p.owner_sees_participation_stats)) {
         fields.owner_sees_participation_stats = !!req.p.owner_sees_participation_stats;
@@ -9249,8 +8577,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 500, "polis_err_get_participant_metadata_choices", err);
     });
   }
-
-
   function handle_GET_metadata_answers(req, res) {
     let zid = req.p.zid;
     let zinvite = req.p.zinvite;
@@ -9293,8 +8619,6 @@ Email verified! You can close this tab or hit the back button.
       doneChecking(false);
     }
   }
-
-
   function handle_GET_metadata(req, res) {
     let zid = req.p.zid;
     let zinvite = req.p.zinvite;
@@ -9369,8 +8693,6 @@ Email verified! You can close this tab or hit the back button.
       doneChecking(false);
     }
   }
-
-
   function getConversationHasMetadata(zid) {
     return new Promise(function(resolve, reject) {
       pgQuery_readOnly('SELECT * from participant_metadata_questions where zid = ($1)', [zid], function(err, metadataResults) {
@@ -9454,8 +8776,6 @@ Email verified! You can close this tab or hit the back button.
     let context = req.p.context;
     // let limit = req.p.limit;
     winston.log("info", "thecontext", context);
-
-
     // this statement is currently a subset of the next one
     // let zidListQuery = "select zid from page_ids where site_id = (select site_id from users where uid = ($1))";
 
@@ -9466,8 +8786,6 @@ Email verified! You can close this tab or hit the back button.
       zidListQuery += " UNION ALL select zid, 2 as type from participants where uid = ($1)"; // using UNION ALL instead of UNION to ensure we get all the 1's and 2's (I'm not sure if we can guarantee the 2's won't clobber some 1's if we use UNION)
     }
     zidListQuery += ";";
-
-
     pgQuery_readOnly(zidListQuery, [uid], function(err, results) {
       if (err) {
         fail(res, 500, "polis_err_get_conversations_participated_in", err);
@@ -9503,8 +8821,6 @@ Email verified! You can close this tab or hit the back button.
           orClauses = orClauses.or(sql_conversations.zid.in(participantInOrSiteAdminOf));
         }
       }
-
-
       query = query.where(orClauses);
       if (!_.isUndefined(req.p.course_invite)) {
         query = query.and(sql_conversations.course_id.equals(req.p.course_id));
@@ -9537,8 +8853,6 @@ Email verified! You can close this tab or hit the back button.
           return;
         }
         let data = result.rows || [];
-
-
         addConversationIds(data).then(function(data) {
           let suurlsPromise;
           if (xid) {
@@ -9641,8 +8955,6 @@ Email verified! You can close this tab or hit the back button.
       return pgQueryP("insert into reports (zid, report_id) values ($1, $2);", [zid, report_id]);
     });
   }
-
-
   function handle_POST_reports(req, res) {
     let zid = req.p.zid;
     let uid = req.p.uid;
@@ -9658,8 +8970,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 500, "polis_err_post_reports_misc", err);
     });
   }
-
-
   function handle_PUT_reports(req, res) {
     let rid = req.p.rid;
     let uid = req.p.uid;
@@ -9706,8 +9016,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 500, "polis_err_post_reports_misc", err);
     });
   }
-
-
   function handle_GET_reports(req, res) {
     let zid = req.p.zid;
     let rid = req.p.rid;
@@ -9762,8 +9070,6 @@ Email verified! You can close this tab or hit the back button.
       }
     });
   }
-
-
   function decodeParams(encodedStringifiedJson) {
     if (!encodedStringifiedJson.match(/^\/?ep1_/)) {
       throw new Error("wrong encoded params prefix");
@@ -9799,8 +9105,6 @@ Email verified! You can close this tab or hit the back button.
     }
     res.send("https://pol.is/settings/enterprise/" + encodeParams(o));
   }
-
-
   function handle_GET_stripe_account_connect(req, res) {
     var stripe_client_id = process.env.STRIPE_CLIENT_ID;
 
@@ -9811,8 +9115,6 @@ Email verified! You can close this tab or hit the back button.
       "<a href ='" + stripeUrl + "'>Connect Pol.is to Stripe</a>" +
       "</body></html>");
   }
-
-
   function handle_GET_stripe_account_connected_oauth_callback(req, res) {
 
     var code = req.p.code;
@@ -9864,8 +9166,6 @@ Email verified! You can close this tab or hit the back button.
         });
     });
   }
-
-
   function handle_GET_conversations(req, res) {
     let courseIdPromise = Promise.resolve();
     if (req.p.course_invite) {
@@ -9930,8 +9230,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 500, "polis_err_post_contexts_check_misc", err);
     });
   }
-
-
   function isUserAllowedToCreateConversations(uid, callback) {
     callback(null, true);
     // pgQuery("select is_owner from users where uid = ($1);", [uid], function(err, results) {
@@ -9955,8 +9253,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 500, "polis_err_reserve_conversation_id", err);
     });
   }
-
-
   function handle_POST_conversations(req, res) {
 
     let xidStuffReady = Promise.resolve();
@@ -9975,8 +9271,6 @@ Email verified! You can close this tab or hit the back button.
           fail(res, 403, "polis_err_add_conversation_not_enabled", new Error("polis_err_add_conversation_not_enabled"));
           return;
         }
-
-
         let q = sql_conversations.insert({
           owner: req.p.uid, // creator
           org_id: req.p.org_id || req.p.uid, // assume the owner is the creator if there's no separate owner specified (
@@ -10009,7 +9303,7 @@ Email verified! You can close this tab or hit the back button.
           let zid = result && result.rows && result.rows[0] && result.rows[0].zid;
 
           const zinvitePromise = req.p.conversation_id ?
-            getZidFromConversationId(req.p.conversation_id).then((zid) => {
+            Conversation.getZidFromConversationId(req.p.conversation_id).then((zid) => {
               return zid === 0 ? req.p.conversation_id : null;
             }) :
             generateAndRegisterZinvite(zid, generateShortUrl);
@@ -10033,8 +9327,6 @@ Email verified! You can close this tab or hit the back button.
       fail(res, 500, "polis_err_conversation_create", err);
     }); // end xidStuffReady
   } // end post conversations
-
-
 
   function handle_POST_query_participants_by_metadata(req, res) {
     let uid = req.p.uid;
@@ -10068,8 +9360,6 @@ Email verified! You can close this tab or hit the back button.
 
     isOwnerOrParticipant(zid, uid, doneChecking);
   }
-
-
   function handle_POST_sendCreatedLinkToEmail(req, res) {
     winston.log("info", req.p);
     pgQuery_readOnly("SELECT * FROM users WHERE uid = $1", [req.p.uid], function(err, results) {
@@ -10149,8 +9439,6 @@ Thanks for using Polis!
       fromAddress,
       body,
     });
-
-
     sendTextEmail(
       fromAddress,
       email,
@@ -10210,8 +9498,6 @@ Thanks for using Polis!
       fail(res, 500, "polis_err_twitter_auth_01", err);
     });
   }
-
-
   function getTwitterAccessToken(body) {
     let oauth = new OAuth.OAuth(
       'https://api.twitter.com/oauth/request_token', // null
@@ -10247,8 +9533,6 @@ Thanks for using Polis!
   let twitterUserInfoCache = new LruCache({
     max: 10000,
   });
-
-
   function getTwitterUserInfo(o, useCache) {
     console.log("getTwitterUserInfo", o);
 
@@ -10305,8 +9589,6 @@ Thanks for using Polis!
       );
     });
   }
-
-
 
   function getTwitterTweetById(twitter_tweet_id) {
     let oauth = new OAuth.OAuth(
@@ -10371,16 +9653,12 @@ Thanks for using Polis!
   //   });
   // }
 
-
-
   // Certain twitter ids may be suspended.
   // Twitter will error if we request info on them.
   //  so keep a list of these for as long as the server is running,
   //  so we don't repeat requests for them.
   // This is probably not optimal, but is pretty easy.
   let suspendedOrPotentiallyProblematicTwitterIds = [];
-
-
   function getTwitterUserInfoBulk(list_of_twitter_user_id) {
     list_of_twitter_user_id = list_of_twitter_user_id || [];
     let oauth = new OAuth.OAuth(
@@ -10423,8 +9701,6 @@ Thanks for using Polis!
       );
     });
   }
-
-
   function switchToUser(req, res, uid) {
     return new Promise(function(resolve, reject) {
       startSession(uid, function(errSess, token) {
@@ -10440,8 +9716,6 @@ Thanks for using Polis!
       });
     });
   }
-
-
   // retry, resolving with first success, or rejecting with final error
   function retryFunctionWithPromise(f, numTries) {
     return new Promise(function(resolve, reject) {
@@ -10460,8 +9734,6 @@ Thanks for using Polis!
       });
     });
   }
-
-
   function updateSomeTwitterUsers() {
     return pgQueryP_readOnly("select uid, twitter_user_id from twitter_users where modified < (now_as_millis() - 30*60*1000) order by modified desc limit 100;").then(function(results) {
       let twitter_user_ids = _.pluck(results, "twitter_user_id");
@@ -10513,13 +9785,9 @@ Thanks for using Polis!
   // Ensure we don't call this more than 60 times in each 15 minute window (across all of our servers/use-cases)
   setInterval(updateSomeTwitterUsers, 1 * 60 * 1000);
   updateSomeTwitterUsers();
-
-
   // getTwitterTweetById("627881237101846528").then(function(tweet) {
   //     console.dir(tweet);
   // });
-
-
   function createUserFromTwitterInfo(o) {
     return createDummyUser().then(function(uid) {
       return getAndInsertTwitterUser(o, uid).then(function(result) {
@@ -10533,8 +9801,6 @@ Thanks for using Polis!
       });
     });
   }
-
-
   function prepForQuoteWithTwitterUser(quote_twitter_screen_name, zid) {
     let query = pgQueryP("select * from twitter_users where screen_name = ($1);", [quote_twitter_screen_name]);
     return addParticipantByTwitterUserId(query, {
@@ -10552,8 +9818,6 @@ Thanks for using Polis!
       }, zid, tweet);
     });
   }
-
-
   function addParticipantByTwitterUserId(query, o, zid, tweet) {
     function addParticipantAndFinish(uid, twitterUser, tweet) {
       return addParticipant(zid, uid).then(function(rows) {
@@ -10615,8 +9879,6 @@ Thanks for using Polis!
       return pgQueryP("INSERT INTO participants (pid, zid, uid, created) VALUES (NULL, $1, $2, default) RETURNING *;", [zid, uid]);
     });
   }
-
-
   function getAndInsertTwitterUser(o, uid) {
     return getTwitterUserInfo(o, false).then(function(u) {
       u = JSON.parse(u)[0];
@@ -10656,8 +9918,6 @@ Thanks for using Polis!
         });
     });
   }
-
-
 
   function handle_GET_twitter_oauth_callback(req, res) {
     let uid = req.p.uid;
@@ -10803,14 +10063,6 @@ Thanks for using Polis!
     });
   }
 
-  function getTwitterInfo(uids) {
-    return pgQueryP_readOnly("select * from twitter_users where uid in ($1);", uids);
-  }
-
-  function getFacebookInfo(uids) {
-    return pgQueryP_readOnly("select * from facebook_users where uid in ($1);", uids);
-  }
-
   function getSocialParticipantsForMod_timed() {
     let start = Date.now();
     return getSocialParticipantsForMod.apply(null, arguments).then(function(results) {
@@ -10940,8 +10192,6 @@ Thanks for using Polis!
       "select uid, 999 as priority from p where mod >= 2) as asdfasdjfioasjdfoi " +
       // "inner join p on asdfasdjfioasjdfoi.uid = p.uid " +
       "group by asdfasdjfioasjdfoi.uid order by priority desc, asdfasdjfioasjdfoi.uid asc), " +
-
-
       // "mod_pptpts2 as (select fjoisjdfio.uid, max(fjoisjdfio.priority) as priority "+
       //     "from ( " +
       //         "select * from pptpts " +
@@ -11024,37 +10274,18 @@ Thanks for using Polis!
   //   return p;
   // }
 
-  function getSocialInfoForUsers(uids, zid) {
-    uids = _.uniq(uids);
-    uids.forEach(function(uid) {
-      if (!_.isNumber(uid)) {
-        throw "polis_err_123123_invalid_uid got:" + uid;
-      }
-    });
-    if (!uids.length) {
-      return Promise.resolve([]);
-    }
-    let uidString = uids.join(",");
-    return pgQueryP_metered_readOnly("getSocialInfoForUsers", "with "+
-      "x as (select * from xids where uid in (" + uidString + ") and owner  in (select org_id from conversations where zid = ($1))), "+
-      "fb as (select * from facebook_users where uid in (" + uidString + ")), "+
-      "tw as (select * from twitter_users where uid in (" + uidString + ")), "+
-      "foo as (select *, coalesce(fb.uid, tw.uid) as foouid from fb full outer join tw on tw.uid = fb.uid) "+
-      "select *, coalesce(foo.foouid, x.uid) as uid from foo full outer join x on x.uid = foo.foouid;", [zid]);
-  }
+  const getSocialInfoForUsers = User.getSocialInfoForUsers;
 
   function updateVoteCount(zid, pid) {
     // return pgQueryP("update participants set vote_count = vote_count + 1 where zid = ($1) and pid = ($2);",[zid, pid]);
     return pgQueryP("update participants set vote_count = (select count(*) from votes where zid = ($1) and pid = ($2)) where zid = ($1) and pid = ($2)", [zid, pid]);
   }
 
-
   // zid_pid => "math_tick:ppaddddaadadaduuuuuuuuuuuuuuuuu"; // not using objects to save some ram
   // TODO consider "p2a24a2dadadu15" format
   let votesForZidPidCache = new LruCache({
     max: 5000,
   });
-
 
   function getVotesForZidPidWithTimestampCheck(zid, pid, math_tick) {
     let key = zid + "_" + pid;
@@ -11069,15 +10300,11 @@ Thanks for using Polis!
     }
     return null;
   }
-
-
   function cacheVotesForZidPidWithTimestamp(zid, pid, math_tick, votes) {
     let key = zid + "_" + pid;
     let val = math_tick + ":" + votes;
     votesForZidPidCache.set(key, val);
   }
-
-
   // returns {pid -> "adadddadpupuuuuuuuu"}
   function getVotesForZidPidsWithTimestampCheck(zid, pids, math_tick) {
     let cachedVotes = pids.map(function(pid) {
@@ -11115,8 +10342,6 @@ Thanks for using Polis!
       return Object.assign(newPidToVotes, cachedPidToVotes);
     });
   }
-
-
   function getVotesForPids(zid, pids) {
     if (pids.length === 0) {
       return Promise.resolve([]);
@@ -11128,8 +10353,6 @@ Thanks for using Polis!
       return votesRows;
     });
   }
-
-
 
   function createEmptyVoteVector(greatestTid) {
     let a = [];
@@ -11175,8 +10398,6 @@ Thanks for using Polis!
     });
     return vectors2;
   }
-
-
   function getLocationsForParticipants(zid) {
     return pgQueryP_readOnly("select * from participant_locations where zid = ($1);", [zid]);
   }
@@ -11218,8 +10439,6 @@ Thanks for using Polis!
       return pids;
     });
   }
-
-
 
   function geoCodeWithGoogleApi(locationString) {
     let googleApiKey = process.env.GOOGLE_API_KEY;
@@ -11267,8 +10486,6 @@ Thanks for using Polis!
       }
     });
   }
-
-
   let twitterShareCountCache = LruCache({
     maxAge: 1000 * 60 * 30, // 30 minutes
     max: 999,
@@ -11315,8 +10532,6 @@ Thanks for using Polis!
       return shares;
     });
   }
-
-
   function getParticipantDemographicsForConversation(zid) {
     return pgQueryP("select * from demographic_data left join participants on participants.uid = demographic_data.uid where zid = ($1);", [zid]);
   }
@@ -11324,8 +10539,6 @@ Thanks for using Polis!
   function getParticipantVotesForCommentsFlaggedWith_is_meta(zid) {
     return pgQueryP("select tid, pid, vote from votes_latest_unique where zid = ($1) and tid in (select tid from comments where zid = ($1) and is_meta = true)", [zid]);
   }
-
-
   function handle_GET_groupDemographics(req, res) {
     let zid = req.p.zid;
     Promise.all([
@@ -11388,8 +10601,6 @@ Thanks for using Polis!
           break;
         }
       }
-
-
       meta = _.indexBy(meta, 'pid');
       let pidToMetaVotes = _.groupBy(metaVotes, 'pid');
 
@@ -11479,8 +10690,6 @@ Thanks for using Polis!
         s.birth_year = s.birth_year / s.birth_year_count;
       }
 
-
-
       res.json(groupStats);
 
     }).catch((err) => {
@@ -11538,8 +10747,6 @@ Thanks for using Polis!
       fail(res, 500, "polis_err_locations_01", err);
     });
   }
-
-
   function removeNullOrUndefinedProperties(o) {
     for (var k in o) {
       let v = o[k];
@@ -11601,8 +10808,6 @@ Thanks for using Polis!
     }
     return x;
   }
-
-
   function handle_PUT_ptptois(req, res) {
     let zid = req.p.zid;
     let uid = req.p.uid;
@@ -11620,8 +10825,6 @@ Thanks for using Polis!
       fail(res, 500, "polis_err_ptptoi_misc_234", err);
     });
   }
-
-
   function handle_GET_ptptois(req, res) {
     let zid = req.p.zid;
     let mod = req.p.mod;
@@ -11668,8 +10871,6 @@ Thanks for using Polis!
       fail(res, 500, "polis_err_famous_proj_get1", err);
     });
   }
-
-
 
   function doFamousQuery(o, req) {
     let uid = o.uid;
@@ -11727,8 +10928,6 @@ Thanks for using Polis!
         });
       });
     }
-
-
     return Promise.all([getConversationInfo(zid), getAuthorUidsOfFeaturedComments()]).then(function(a) {
       let conv = a[0];
       let authorUids = a[1];
@@ -11785,8 +10984,6 @@ Thanks for using Polis!
         //     }
         //     return false;
         // }
-
-
         // TODO There are issues with this:
         //   really, the data should all be merged first, then the list should be truncated to the correct number.
         // ALSO, we could return data on everyone who might appear in the list view, and add an "importance" score to help determine who to show in the vis at various screen sizes. (a client determination)
@@ -11940,8 +11137,6 @@ Thanks for using Polis!
       });
     });
   }
-
-
 /*
 
 CREATE TABLE slack_users (
@@ -11958,8 +11153,6 @@ CREATE TABLE slack_user_invites (
     created BIGINT DEFAULT now_as_millis()
 );
 */
-
-
 
   function handle_GET_slack_login(req, res) {
 
@@ -12044,8 +11237,6 @@ CREATE TABLE slack_user_invites (
       });
     });
   }
-
-
   function handle_POST_slack_interactive_messages(req, res) {
     const payload = JSON.parse(req.p.payload);
 
@@ -12065,8 +11256,6 @@ CREATE TABLE slack_user_invites (
 
     console.dir(response_url);
     console.dir(payload);
-
-
 
     postMessageUsingHttp({
       channel: channel.id,
@@ -12158,8 +11347,6 @@ CREATE TABLE slack_user_invites (
   //   .pipe(res);
 
   // }
-
-
   function handle_GET_einvites(req, res) {
     let einvite = req.p.einvite;
 
@@ -12173,8 +11360,6 @@ CREATE TABLE slack_user_invites (
       fail(res, 500, "polis_err_fetching_einvite", err);
     });
   }
-
-
   function handle_POST_contributors(req, res) {
     const uid = req.p.uid || null;
     const agreement_version = req.p.agreement_version;
@@ -12235,13 +11420,9 @@ CREATE TABLE slack_user_invites (
       fail(res, 500, "polis_err_POST_waitinglist", err);
     });
   }
-
-
   function generateSingleUseUrl(req, conversation_id, suzinvite) {
     return getServerNameWithProtocol(req) + "/ot/" + conversation_id + "/" + suzinvite;
   }
-
-
   function buildConversationUrl(req, zinvite) {
     return getServerNameWithProtocol(req) + "/" + zinvite;
   }
@@ -12263,8 +11444,6 @@ CREATE TABLE slack_user_invites (
       return buildConversationUrl(req, zinvite);
     });
   }
-
-
   function createOneSuzinvite(xid, zid, owner, generateSingleUseUrl) {
     return generateSUZinvites(1).then(function(suzinviteArray) {
       let suzinvite = suzinviteArray[0];
@@ -12287,8 +11466,6 @@ CREATE TABLE slack_user_invites (
     });
   }
 
-
-
   function renderLtiLinkagePage(req, res, afterJoinRedirectUrl) {
     let context_id = req.p.context_id;
     let user_id = req.p.user_id;
@@ -12296,8 +11473,6 @@ CREATE TABLE slack_user_invites (
     let tool_consumer_instance_guid = req.p.tool_consumer_instance_guid;
 
     let greeting = '';
-
-
     // TODO If we're doing this basic form, we can't just return json from the /login call
 
     let form1 = '' +
@@ -12359,8 +11534,6 @@ CREATE TABLE slack_user_invites (
       '<button type="submit" class="Btn Btn-primary">Sign In</button>' +
       '</div>' +
       '</form></p>';
-
-
     res.set({
       'Content-Type': 'text/html',
     });
@@ -12383,8 +11556,6 @@ CREATE TABLE slack_user_invites (
 
     res.status(200).send(html);
   }
-
-
   // team meetings - schedule with others, smart converence room
   // or redirect tool
   // students already pay an online fee
@@ -12479,7 +11650,7 @@ CREATE TABLE slack_user_invites (
             //     // you're good!
             // } else {
             //     if (you paid) {
-            renderLtiLinkageSuccessPage(req, res, {
+            User.renderLtiLinkageSuccessPage(req, res, {
               context_id: context_id,
               // user_image: userForLtiUserId.user_image,
               email: userForLtiUserId.email,
@@ -12510,8 +11681,6 @@ CREATE TABLE slack_user_invites (
     });
   } // end /api/v3/LTI/setup_assignment
 
-
-
   // function handle_POST_lti_canvas_nav(req, res) {
   //     winston.log("info",req);
   //     let roles = req.p.roles;
@@ -12530,8 +11699,6 @@ CREATE TABLE slack_user_invites (
 
   //     // Check if linked to this uid.
   //     pgQueryP("select * from lti_users left join users on lti_users.uid = users.uid where lti_users.lti_user_id = ($1) and lti_users.tool_consumer_instance_guid = ($2);", [user_id, req.p.tool_consumer_instance_guid]).then(function(rows) {
-
-
   //         let userForLtiUserId = null;
   //         if (rows.length) {
   //             userForLtiUserId = rows[0];
@@ -12544,23 +11711,17 @@ CREATE TABLE slack_user_invites (
   //             //     // you're good!
   //             // } else {
   //             //     if (you paid) {
-
-
   //                     // renderLtiLinkageSuccessPage(req, res, {
   //                     //     context_id: context_id,
   //                     //     // user_image: userForLtiUserId.user_image,
   //                     //     email: userForLtiUserId.email,
   //                     // });
-
-
   //                     let inboxLaunchParams = encodeParams({
   //                         context: context_id, // we're using the LTI context_id as a polis conversation context. scope the inbox to the course
   //                         xPolisLti: createPolisLtiToken(req.p.tool_consumer_instance_guid, req.p.user_id),  // x-polis-lti header
   //                         // TODO add token
   //                     });
   //                     res.redirect("https://preprod.pol.is/inbox/" + inboxLaunchParams);
-
-
 
   //                 // } else { // you (student) have not yet paid
   //                 //     // gotta pay
@@ -12582,8 +11743,6 @@ CREATE TABLE slack_user_invites (
   //     });
 
   // } // end /api/v3/LTI/canvas_nav
-
-
 
   function addCanvasAssignmentConversationInfoIfNeeded(zid, tool_consumer_instance_guid, lti_context_id, custom_canvas_assignment_id) {
     return getCanvasAssignmentInfo(tool_consumer_instance_guid, lti_context_id, custom_canvas_assignment_id).then(function(rows) {
@@ -12647,8 +11806,6 @@ CREATE TABLE slack_user_invites (
       tool_consumer_instance_guid,
     ]);
   }
-
-
   function handle_POST_lti_conversation_assignment(req, res) {
     let roles = req.p.roles;
     let isInstructor = /[iI]nstructor/.exec(roles); // others: Learner
@@ -12682,8 +11839,6 @@ CREATE TABLE slack_user_invites (
         winston.log("info", err);
       });
     }
-
-
     function constructConversationUrl(zid) {
       // sweet! the instructor has created the conversation. send students here. (instructors too)
       return getZinvite(zid).then(function(zinvite) {
@@ -12800,13 +11955,9 @@ CREATE TABLE slack_user_invites (
     // so, on conversation close... we should keep "canvas_assignment_conversation_info": a many-to-many mapping of {zid <=> (tool_consumer_instance_guid, lti_context_id, custom_canvas_assignment_id)}
     // so iterate over all triples for the zid, and find the corresponding callback record, and make a signed request for each student's record.
     // Note that if the student somehow joined the conversation, but not through canvas, then they can't get credit.
-
-
     // wait! how do we know what the conversation should have for topic / description?
 
   }
-
-
 
   function handle_GET_setup_assignment_xml(req, res) {
     let xml = '' +
@@ -12852,8 +12003,6 @@ CREATE TABLE slack_user_invites (
       //     '<lticm:property name="enabled">true</lticm:property>' +
       // '</lticm:options>' +
 
-
-
       '</blti:extensions>' +
 
       '<cartridge_bundle identifierref="BLTI001_Bundle"/>' +
@@ -12863,8 +12012,6 @@ CREATE TABLE slack_user_invites (
     res.set('Content-Type', 'text/xml');
     res.status(200).send(xml);
   }
-
-
   function handle_GET_conversation_assigmnent_xml(req, res) {
     let serverName = getServerNameWithProtocol(req);
 
@@ -12911,8 +12058,6 @@ CREATE TABLE slack_user_invites (
     res.set('Content-Type', 'text/xml');
     res.status(200).send(xml);
   }
-
-
   function handle_GET_canvas_app_instructions_png(req, res) {
     let path = "/landerImages/";
     if (/Android/.exec(req.headers['user-agent'])) {
@@ -12928,14 +12073,10 @@ CREATE TABLE slack_user_invites (
     doFetch(req, res);
   }
 
-
-
   // function handle_POST_users_invite(req, res) {
   //     let owner = req.p.uid;
   //     let xids = req.p.xids;
   //     let zid = req.p.zid;
-
-
   //     // generate some tokens
   //     // add them to a table paired with user_ids
   //     // return URLs with those.
@@ -12970,8 +12111,6 @@ CREATE TABLE slack_user_invites (
   //         fail(res, 500, "polis_err_generating_single_use_invites", err);
   //     });
   // }
-
-
 
   function hangle_GET_testConnection(req, res) {
     res.status(200).json({
@@ -13067,8 +12206,6 @@ CREATE TABLE slack_user_invites (
     });
   }
 
-
-
   function initializeImplicitConversation(site_id, page_id, o) {
 
     // find the user with that site_id.. wow, that will be a big index..
@@ -13079,8 +12216,6 @@ CREATE TABLE slack_user_invites (
         throw new Error("polis_err_bad_site_id");
       }
       return new Promise(function(resolve, reject) {
-
-
         let uid = rows[0].uid;
         //    create a conversation for the owner we got,
         let generateShortUrl = false;
@@ -13196,7 +12331,7 @@ CREATE TABLE slack_user_invites (
 
   function doGetConversationPreloadInfo(conversation_id) {
     // return Promise.resolve({});
-    return getZidFromConversationId(conversation_id).then(function(zid) {
+    return Conversation.getZidFromConversationId(conversation_id).then(function(zid) {
       return Promise.all([
         getConversationInfo(zid),
       ]);
@@ -13240,8 +12375,6 @@ CREATE TABLE slack_user_invites (
       fail(res, 500, "polis_err_get_conversation_preload_info", err);
     });
   }
-
-
 
   // NOTE: this isn't optimal
   // rather than code for a new URL scheme for implicit conversations,
@@ -13292,8 +12425,6 @@ CREATE TABLE slack_user_invites (
       o.bgcolor = req.p.bg_white ? "#fff" : null;
     }
     o.socialbtn_type = req.p.show_share ? 1 : 0;
-
-
     // Set stuff in cookies to be retrieved when POST participants is called.
     let setOnPolisDomain = !domainOverride;
     let origin = req.headers.origin || "";
@@ -13436,8 +12567,6 @@ CREATE TABLE slack_user_invites (
     //     // });
     //     getStaticFile("./unsupportedBrowser.html", res);
     // } else {
-
-
     let port = process.env.STATIC_FILES_PORT;
     // set the host header too, since S3 will look at that (or the routing proxy will patch up the request.. not sure which)
     req.headers.host = hostname;
@@ -13583,8 +12712,6 @@ CREATE TABLE slack_user_invites (
   let hostname = process.env.STATIC_FILES_HOST;
   let portForParticipationFiles = process.env.STATIC_FILES_PORT;
   let portForAdminFiles = process.env.STATIC_FILES_ADMINDASH_PORT;
-
-
   let fetchUnsupportedBrowserPage = makeFileFetcher(hostname, portForParticipationFiles, "/unsupportedBrowser.html", {
     'Content-Type': "text/html",
   });
@@ -13633,13 +12760,9 @@ CREATE TABLE slack_user_invites (
 
     return fetchIndex(req, res, {}, port);
   }
-
-
   function ifDefinedFirstElseSecond(first, second) {
     return _.isUndefined(first) ? second : first;
   }
-
-
   let fetch404Page = makeFileFetcher(hostname, portForAdminFiles, "/404.html", {
     'Content-Type': "text/html",
   });
@@ -13682,16 +12805,12 @@ CREATE TABLE slack_user_invites (
       // fail(res, 500, "polis_err_fetching_conversation_info2", err);
     });
   }
-
-
   let fetchIndexForAdminPage = makeFileFetcher(hostname, portForAdminFiles, "/index_admin.html", {
     'Content-Type': "text/html",
   });
   let fetchIndexForReportPage = makeFileFetcher(hostname, portForAdminFiles, "/index_report.html", {
     'Content-Type': "text/html",
   });
-
-
 
   function handle_GET_iip_conversation(req, res) {
     let conversation_id = req.params.conversation_id;
@@ -13700,8 +12819,6 @@ CREATE TABLE slack_user_invites (
     });
     res.send("<a href='https://pol.is/" + conversation_id + "' target='_blank'>" + conversation_id + "</a>");
   }
-
-
   function handle_GET_iim_conversation(req, res) {
     let zid = req.p.zid;
     let conversation_id = req.params.conversation_id;
@@ -13718,8 +12835,6 @@ CREATE TABLE slack_user_invites (
       fail(res, 500, "polis_err_fetching_conversation_info", err);
     });
   }
-
-
 
   function handle_GET_twitter_image(req, res) {
     console.log("handle_GET_twitter_image", req.p.id);
@@ -13764,8 +12879,6 @@ CREATE TABLE slack_user_invites (
       res.status(500).end();
     });
   }
-
-
   let handle_GET_conditionalIndexFetcher = (function() {
     return function(req, res) {
       if (hasAuthToken(req)) {
@@ -13785,8 +12898,6 @@ CREATE TABLE slack_user_invites (
       }
     };
   }());
-
-
 
   function handle_GET_localFile_dev_only(req, res) {
     let filename = String(req.path).split("/");
@@ -13869,8 +12980,6 @@ CREATE TABLE slack_user_invites (
       addInRamMetric(path, time);
     }
   });
-
-
   console.log('end initializePolisHelpers');
 
   return {
@@ -13908,14 +13017,10 @@ CREATE TABLE slack_user_invites (
     winston,
     writeDefaultHead,
     yell,
-
-
     middleware_check_if_options,
     middleware_log_middleware_errors,
     middleware_log_request_body,
     middleware_responseTime_start,
-
-
     // handlers
     handle_DELETE_metadata_answers,
     handle_DELETE_metadata_questions,
@@ -14046,17 +13151,12 @@ CREATE TABLE slack_user_invites (
   };
 
 } // End of initializePolisHelpers
-
-
 // debugging
 //let ph = initializePolisHelpers()
 
 //if (false) {
   //let nextP = ph.getNextPrioritizedComment(17794, 100, [], true);
 //};
-
-
-
 
 module.exports = {
   initializePolisHelpers,
